requirements.txt
torch>=1.12.0
torchvision>=0.13.0
Pillow>=8.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
matplotlib>=3.5.0